test(framework): apply self-review findings on reflection removal sweep
